Moved Di Related  to Di Package and VM Factory to Factory
