+++ domain/src/main/java/com/practical/domain/ResultState.kt
@@ -1,0 +1,7 @@
+package com.practical.domain
+
+sealed class ResultState< out T> {
+    data object Loading : ResultState<Nothing>()
+    data class Success<T>(val data: T) : ResultState<T>()
+    data class Error(val exception: Throwable) : ResultState<Nothing>()
+}
